requirements.txt
torch==2.0.1 
torchvision==0.15.2
openmim==0.3.9
pandas==1.4.0
albumentations==1.3.1
easyocr==1.7.1
pillow==10.2.0
pytesseract==0.3.10
openai==1.11.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80